sub2_backend/requirements.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.16
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
zipp==1.2.0
